moved window utils
